requirements.txt
bitsandbytes==0.42.0
gradio==4.20.0
scipy==1.12.0
llama_cpp_python>=0.2.44
